Automatically create ADSManagerRequest when saving ADSManagerAdministrator
ADSManagerAdministrator represents an employee of a prefecture, who
can grant permissions to other users to manage the ADS of an
administration depending on this prefecture.

This employee should have the permissions to manage ADS of the
prefecture.

When we edit the list of users of an ADSManagerAdministrator, we now
grant (or revoke) access to manage the ADS of the related prefecture.