ADSManagerAdministrators can manage ADS for all administrations depending on their prefecture
State employee working for the Prefecture of l'Ille-Et-Vilaine for
example can now manage ADS for all administrations under the prefecture
(communes, EPCI, and the prefecture itself).

Before, we provided this feature by creating entries in
ADSManagerRequest. It was not ideal because it pollutes listings and
creates inconsistencies (what happens if you reject or delete the
request of an administrator? he can no longer manage the ADS, but he can
make a request and accept it himself to get the permissions back).

Now, administrators can manage all administrations depending on their
prefecture.