Fix ADS form rendering when invalid
Django forms are a pain to work with dates.

If a form which contains a date is valid, then the date field is a
datetime.date object. But if the form is invalid, then the input's value
is the previously submitted value, which is a string.

By calling |str_to_date on the field value, you can make sure the field
is always a date object you can call |date on.