Fix unicode error in import_last_update_file_from_paris
This is a tricky bug. csv.DictReader takes an iterable as argument. If
this is a file handle, then the iterator attempts to read 2^16 bytes at
a time. If the last byte of the buffer is in the middle of a utf-8
sequence, an exception is raised.

I couldn't find a way reproduce the bug easily, so to avoid the
buffering I simply read the whole file in memory.