Test setup htmx, with a button on ads detail page to load the faq
WIP WIP WIP

One of the issue with htmx that I don't konw how to solve yet is how to
keep data when the form is replaced.

For example, user sets ADS creation date to > 2014, starts filling
fields, then changes the date for < 2014. How to keep the data
previously entered? We don't want to discard everything.