fix import ADS of Paris
The CSV file sent by Paris contains a field with the attribution date,
but not the creation date.

If "type" is "Gratuite non cessible", it's a new ADS and the date should
be considered as the creation date of the ADS.

If the type is anything else, it's an old ADS, this date is the
attribution date, and we don't have the creation date of the ADS, so we
hardcode it.