Fix stats showing the number of complete ADS
To know whether an ADS is complete (ie. has all the important details
defined) or not, we use to check if there was an entry in the table
ADSUpdateLog.

There were two problems with this:
- there might be more than one complete ADSUpdateLog, and we want to
  know whether the ADS is complete or not, not to know how many
  "complete" logs exist
- and we only want to get the latest log. If the log was complete in the
  past but it's no longer the case now, we should not account the entry
  as complete.

There are two ways the "complete" status is computed: either by using
the ADSManager object, or through the backward relationship in
ADSManagerAdministrator. Hence the huge diff.